docs: Update

+++ model/sleep/sleep.go
@@ -38,7 +38,7 @@ type Sleep struct {
 	UserID     int64        `xorm:"'user_id' index" json:"user_id"`
 	StartTime  time.Time    `xorm:"start_time" json:"start_time"`
 	EndTime    time.Time    `xorm:"end_time" json:"end_time"`
-	Duration   int          `xorm:"duration" json:"duration"`
+	Duration   float64      `xorm:"duration" json:"duration"`
 	Quality    SleepQuality `xorm:"quality" json:"quality"`
 	RecordedAt time.Time    `xorm:"recorded_at" json:"recorded_at"`
 }
